--- a/bench_test.go
+++ b/bench_test.go
@@ -14,7 +14,7 @@ func benchPayload() []byte {
 		p = append(p, 'i')
 	}
 	return []byte(`{
-""input":"` + string(p) + `
+"input":"` + string(p) + `
 "}`)
 }