--- a/raw.go
+++ b/raw.go
@@ -1,8 +1,28 @@
 package protocol
 
 import (
+	"bytes"
 	"errors"
+	"sync"
+
+	"github.com/segmentio/encoding/json"
 )
+
+var bufferPool = sync.Pool{
+	// New is called when a new instance is needed
+	New: func() interface{} {
+		return new(bytes.Buffer)
+	},
+}
+
+func getBuffer() *bytes.Buffer {
+	return bufferPool.Get().(*bytes.Buffer)
+}
+
+func putBuffer(buf *bytes.Buffer) {
+	buf.Reset()
+	bufferPool.Put(buf)
+}
 
 // Raw type used by Centrifuge protocol as a type for fields in structs which
 // value we want to stay untouched. For example custom application specific JSON
@@ -15,7 +35,14 @@ func (r Raw) MarshalJSON() ([]byte, error) {
 	if r == nil {
 		return []byte("null"), nil
 	}
-	return r, nil
+	buf := getBuffer()
+	err := json.Compact(buf, r)
+	if err != nil {
+		return nil, err
+	}
+	res := append([]byte(nil), buf.Bytes()...)
+	putBuffer(buf)
+	return res, nil
 }
 
 // UnmarshalJSON sets *r to a copy of data.