… bpool_test.go
… encode_helpers_test.go
… encode_test.go